+++ modules/core/arrow-instances-data/src/main/kotlin/arrow/instances/kleisli.kt
@@ -103,29 +103,6 @@ interface KleisliMonadErrorInstance<F, D, E> : MonadError<KleisliPartialOf<F, D>
   override fun FFF(): MonadError<F, E>
 }
 
-@extension
-interface KleisliBracketInstance<F, R, E> : Bracket<KleisliPartialOf<F, R>, E>, KleisliMonadErrorInstance<F, R, E> {
-
-  override fun FFF(): Bracket<F, E>
-
-  override fun <A, B> Kind<KleisliPartialOf<F, R>, A>.bracketCase(
-    use: (A) -> Kind<KleisliPartialOf<F, R>, B>,
-    release: (A, ExitCase<E>) -> Kind<KleisliPartialOf<F, R>, Unit>
-  ): Kleisli<F, R, B> =
-    FFF().run {
-      Kleisli { r ->
-        this@bracketCase.fix().run(r).bracketCase({ a ->
-          use(a).fix().run(r)
-        }, { a, br ->
-          release(a, br).fix().run(r)
-        })
-      }
-    }
-
-  override fun <A> Kind<KleisliPartialOf<F, R>, A>.uncancelable(): Kleisli<F, R, A> =
-    Kleisli { r -> FFF().run { this@uncancelable.fix().run(r).uncancelable() } }
-}
-
 /**
  * Alias for [Kleisli] for [Id]
  */
